Battle: Spilt class600 to normal and test code #1656
Moved new class600 code to BattlePlayerClass600Test.cs
Reverted BattlePlayerClass600.cs
Updated BattlePlayerClassManager.cs
