+++ Assets/QuantumUser/Simulation/Battle/Scripts/Player/PlayerClasses/BattlePlayerClass600Test.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cc2e02925162efd4eab97a6fd92e36d0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
